+++ packages/interfaces/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- removed an unused key from `parcel` application json schema
+
 ## [0.4.0] - 2022-11-26
 
 ### Fixed